requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
asyncpg==0.29.0

# AI/ML (from your forecasting service)
openai==1.3.7
prophet==1.1.5
xgboost==2.0.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4

# HTTP client for forecast integration
aiohttp==3.9.1
httpx==0.24.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Model persistence (used in your forecasting service)
joblib==1.3.2

# Performance utilities (from your forecasting code)
urllib3<2.0.0
# concurrent-futures is built-in for Python 3.2+, not needed

# CORS support
fastapi-cors==0.0.6
